Support @stream_connection for streamed connections
Summary:
Adds support for a new directive, `stream_connection` which simplifies the usage of streaming with connections. The directive accepts the union of arguments of `connection` and `stream`:

```
directive stream_connection(
    key: String!
    filters: [String]
    handler: String
    label: String!
    initial_count: Int!
    if: Boolean = true
  ) on FIELD
```

The connection args (key/filters/handler) are used as they are today to construct the `handle` applied directly to the field, while the stream args are used to construct a `stream` directive applied to the `edges` field and a `defer` directive to apply to the `pageInfo` field.

This also changes the transform to stop generating a fragment on the connection type (to fetch extra fields on edges/pageInfo) and instead generate two smaller fragments that are appended to the existing edges/pageInfo selections. This ensures that there is only one edges/pageInfo field each to which the stream/defer directives can be applied. The transform is also careful to preserve the original field ordering (ie ordering of edges/pageInfo selections relative to each other and to other selections on the connection).

Reviewed By: kassens

Differential Revision: D14250505

fbshipit-source-id: 97bd1ae51ecb6e1d77604c6d4ead66be8aeaa87f