graphql tags no longer produce implicit objects
Summary:
This changes the babel plugin to no longer produce an implicit object in some cases.

Previously, if a `graphql` tag was not somewhere in an object literal in the parent path, it would produce an object with the name of the fragment after the first `_` (defaulting to `data`) as the key. If these rules sound complicated and confusing to you, you're probably not alone. That's why we're removing this case and simplifying the transform.

Reviewed By: josephsavona, jstejada

Differential Revision: D14247660

fbshipit-source-id: b2567dc7879ed85e560fce595d160d5eacef305d