ConnectionTransform: generate edges/pageInfo w/o aliases
Summary:
D14250505 introduced support for `stream_connection` and changed ConnectionTransform to generate additional selections under the existing edges/pageInfo fields, whereas before it would always generate a single fragment under the connection field. A side-effect of this is that if edges/pageInfo were aliased, the generated fields now show up under the alias rather than as non-aliased sibling selections. We can argue about which behavior is preferable if we were starting from scratch (i'm inclined toward the new behavior) but for backwards compatibility we should  keep the old behavior for now and migrate code if/when we decide to change it.

This diff ensures that we generate edges/pageInfo selections w/o aliases for `connection`, and disallows aliases on those fields when using `stream_connection` just to reduce inconsistency (also see the code comment).

Reviewed By: jstejada, alunyov

Differential Revision: D14288894

fbshipit-source-id: 921ba9f9939dd3f3c16a8552c26c90b919464c1a