Add Rollout for incremental codegen changes
Summary:
This module enables rolling out codegen changes over multiple commits in case there is a maximum commit size.

In the codegen code somewhere:
```
const shouldUseNewVersion = Rollout.check(artifactName);
```
which gives a `boolean` based on some string, the fragment name of the artifact is a good candidate.

In the compiler main, you'd need to set which buckets to enable the codegen for, for example:
```
Rollout.setBuckets([true, false, true, true]) // enables the rollout for 3 out of 4 buckets
```
If the buckets array is split over multiple lines with a comment between, multiple of these buckets can be turned on in parallel and they should merge cleanly.

Reviewed By: jstejada

Differential Revision: D14292228

fbshipit-source-id: f4af74749be1b04e36a7fd0fc42b8b9e0c22bf04