Stricter parsing of @argDefs defaultValues
Summary: Enables stricter parsing of `argumentDefinitions`, disallowing the legacy (and already ignored) list/nonNull properties and validating that the `defaultValue` is valid per the given `type`. For backwards compatibility, the validation of enums is non-strict: instead of requiring a properly-cased enum literal (`WEB`), it also allows string literals and incorrect casing. The raw enum value is preserved so long as it clearly corresponds to a defined enum variant. In the future we can remove this flexible enum parsing - the main motivation here was to ensure that we have stricter parsing of more complex types.

Reviewed By: kassens

Differential Revision: D14411875

fbshipit-source-id: e915a5914fce661760b477b67bbdd90b193a1544