@stream_connection integration w incremental delivery feature flag
Summary: Failing to add an `if` argument on `stream_connection` threw an error, but the construction of the error was failing due to a missing `loc: Location` property. This diff adds the location property where it was missing, and also adds a custom error message to ConnectionTransform to inform the user about the `if` argument requirement.

Reviewed By: kassens

Differential Revision: D14504651

fbshipit-source-id: 1c254439c7ea58aa8410bbc9177fb652f402a72b