Move persisted id generation to RelayCompilerMain (#2641)
Summary:
The current hash that we use for persisted queries (`sourceHash`) does not include the full query text and only the query node.

For example the following query and fragments:

App A:
```graphql
query CardsScreenQuery {
  viewer {
    ...CardsScreen_viewer
  }
}

fragment CardsScreen_viewer on Viewer {
  name
}
```

App B:
```graphql
query CardsScreenQuery {
  viewer {
    ...CardsScreen_viewer
  }
}

fragment CardsScreen_viewer on Viewer {
  logo
}
```

Would generate the same id because the current hash includes only the query part and not the fragment. This means it only hashes:

```graphql
query CardsScreenQuery {
  viewer {
    ...CardsScreen_viewer
  }
}
```

To fix this I added a hash for the whole query text and pass that as the id to `persistQuery`. This makes sure the previous two queries in the example generate a different id.

Note that the ids used to work prior to upgrading to version 2.0 and using `--persisted-output`. Before I used a custom relay-compiler-bin that generated the queries map in `persistQuery`.
Pull Request resolved: https://github.com/facebook/relay/pull/2641

Reviewed By: alunyov

Differential Revision: D14563369

Pulled By: kassens

fbshipit-source-id: 922eae863700903ba73a515ef5d557652c1ee2ac