Unused argDefs causes runtime error 1/2: regression test
Summary:
Adds a runtime unit test to assert that an unused argDef can cause a runtime error and a compiler test fixture to demonstrate why. The specific error occurs when a fragment references a global variable inside `arguments` and the fragment it is passing the argument to doesn't end up using the corresponding argument: in this case, the compiler doesn't see the global variable ever being used, and doesn't check that there is a corresponding query argument definition. Without a query argument definition, Relay doesn't populate a default value for the variable, and it's then null when reader tries to access it.

A follow-up diff will fix the issue.

Reviewed By: jstejada

Differential Revision: D14716989

fbshipit-source-id: 5b572d4d54c9a7367f8de490f68a3734dda6e870