Updated API for MockPayloadGenerator
Summary:
With this diff, I'm adding a logic to handle the metadata, generated by relay_test_operation to MockPayloadGenerator.

Also, after spending some time working with testing tools, I realized that we should reduce the API surface for MockPayloadGenerator to just one method -> generateDataForOperation. As this is the only method I'm actually using in the tests. I don't see the practical use cases for generating data just for fragment selections.

So I in this diff we have,
- Add support for relay_test_operation
- Rename `generateDataForOperation` -> `generate` as this will be the only one method.
- Remove `generateData` (this was working for Reader/Normalization selections)
- Remove `generateVariables` - as we have variables on the Operation.

Reviewed By: josephsavona

Differential Revision: D14791155

fbshipit-source-id: 7f2e17457e7f4bfd7db4b5da59a3de15f3474694