getOrCreateLinkedRecord links to existing records
Summary: Fixes a bug with `getOrCreateLinkedRecord()` when the field had previously been set (so a record w the client id exists) but later unset. Currently we unconditionally attempt to create a record w the appropriate clientID which fails in this case since a record with that ID already exists. This is fixed to first attempt to get a record w the clientID and only create one if it doesn't exist.

Reviewed By: jstejada

Differential Revision: D15351624

fbshipit-source-id: 6d9b0cecf7cc4108cf99a4b9998122d1b5c558a4