Namespace 3D results by document name
Summary:
Data-driven dependencies is currently implemented so that the results of the `module` selection (and the corresponding fragment) are stored at fixed storage keys within the store. In theory this means that two different parts of an app - on the screen or otherwise - could select different 3D modules that gets normalized into the same record, causing a (silent) conflict.

This diff prevents such conflicts by effectively namespacing 3D results based on the name of the document in which they were selected (the query, mutation, subscription, or fragment). This means that the results of a 3D selection can only update if the exact same fragment (document) is re-fetched.

Overview of the changes:

* Adds a `documentName` property on ModuleImport IR/reader/normalization nodes. This is used to compute the alias *and* storage key used to record the results of fetching the `js()` field for the component and operation modules.
* Sets the value of these aliases to the previous fixed name followed by the document name - e.g. `__module_component_<FooFragment_foo>` or `__module_operation_<FooFragment_foo>`.
* Changes the runtime traversals to use the componentKey/operationKey properties of the ModuleImport node instead of the previous static names.
* Finally, the compiler previously emitted the component field outside of the ModuleImport, and the operation field inside of it. This was so that normalizer could eagerly store the matched component even before (possibly asynchronously) normalizing the matched data. This is now changed so both fields are inside the ModuleImport, and normalizer and reader just directly read and set the value of the field (in fact Reader was already doing this, and processing that field twice).

Reviewed By: jstejada

Differential Revision: D15349734

fbshipit-source-id: ae89c5669b2679d29ddec576d26db30790a3cef1