Enable incremental data delivery (w/o feature flag)
Summary: Incremental data delivery has so far been enabled only behind a feature flag. The main restriction is that an `if` argument was required and had to be set to a fixed variable name whose value matched the feature flag value - this gave us a kill-switch during initial development of the feature. We're now ready to experiment with this feature at a larger scale; as such this diff fully enables incremental data delivery (defer/stream/stream_connection), removing the feature flag and associated restrictions.

Reviewed By: jstejada

Differential Revision: D15369789

fbshipit-source-id: b98c75e7566fbd14421c89549ad9ddb535488724