Generate 3D metadata annotations on fragments not operations
Summary:
Changes the way the FB-internal generated module formatter outputs 3D metadata annotations. Before, *normalization* .graphql.js files (for queries/mutations/subscriptions) contained information about 3D dependencies of the component tree that defined the operation. This made it more difficult to allocate file size costs back to the components that actually define the 3D dependencies. This diff changes where metadata is emitted to only include 3D metadata for fragments (including the implicit root fragment of a query/mutation/subscription), so that it's easier to see the 3D dependencies of each part of the app.

Imagine an app w App -> Parent -> Child components, where Child is a 3D dependency defined in Parent. Effectively, the way the metadata was emitted and reported *before* was:

```
- App.js
  - Parent.js
    - Parent_parent.graphql.js
  - AppQuery.graphql.js
    - [3D] Child_child$normalization.graphql.js
    - [3D] Child.js
      - Child_child.graphql.js
```

Note that it appears the 3D dependencies are defined by App.js, but in reality they're defined by Parent.js. With this diff we'll more accurately present the dependencies:

```
- App.js
  - Parent.js
    - Parent_parent.graphql.js
      - [3D] Child_child$normalization.graphql.js
      - [3D] Child.js
        - Child_child.graphql.js
  - AppQuery.graphql.js
```

Note that in both cases the same 3D dependencies roll up to App.js so it *shouldn't* change the total reported size.

Reviewed By: alunyov

Differential Revision: D15405758

fbshipit-source-id: f8c330a697579a18191d6eea8ab37a9030d5a140