QueryExecutor: queue incremental payloads until parent @module resolves
Summary:
Fixes an issue where using defer/stream within a 3D `module` can fail due to a race condition. In order to process an incremental payload, the parent field must have already been processed so that QueryExecutor knows the context (what field handlers to run, the latest version of the parent record, etc). But when defer/stream are nested inside a 3D selection, it's possible that the incremental payload can arrive *before* the 3D data has been normalized.

The fix here is to queue incremental responses if the 'label' is unknown (ie, if the parent is a 3D selection and hasn't been normalized yet).  Then when the parent is normalized, any pending responses are processed.

Doing this required removing an invariant that I had added as a sanity check for invalid label values coming from the server. In a follow-up we may want to add back the sanity checks, but that was more of a convenience during initial development to catch bugs - note that a conforming server wouldn't return back invalid labels. One option would be to attach some metadata to each *root* operation about the set of labels it may provide, which would allow us to be able to distinguish pending labels vs completely invalid labels.

Reviewed By: alunyov

Differential Revision: D15467189

fbshipit-source-id: a1ce552402548fa2f5139f4335d3bf6c8e56345c