Fix for @module on field w concrete type
Summary: Fixes an edge-case with normalization of at-module selections, where when at-module was used on a field with a concrete type no `__typename` selection would be present in the result (since the type is known statically), but the field was being read to produce the ModuleImportPayload value. The fix is to use the typename from the record instead of the result.

Reviewed By: kassens

Differential Revision: D15607341

fbshipit-source-id: ee16e9dc945233b9dbcbfba5adc43a1cdc6c643f