Hack to forward old urls to relay.dev (#2756)
Summary:
I tried adding a forward using GitHubs domain setting, but that clashed
with some other config and triggered an infinite redirect.

This is just a terrible hack in a terrible place to add a forward to
most pages (all that use the footer, this excludes the GraphQL spec
extension pages).
Pull Request resolved: https://github.com/facebook/relay/pull/2756

Reviewed By: alunyov

Differential Revision: D15696715

Pulled By: kassens

fbshipit-source-id: cc74862d922816baec8cd2b2d44108875ff37f76