relay-compiler: add additional defined-variable check for connection keys
Summary: Currently we run the RefineOperationVariables transform on the print context only: it errors if the query references an undefined variable and automatically removes unused variables (which can occur due to things like `arguments` plus include/skip). However, this transform is also (correctly)  run *after* handle fields are stripped - the server doesn't know about them. This makes it difficult to figure out which operations to update if  you add a dynamic connection key variable,  however, since the variable *only* appears in a handle field: the effect is that the transform doesn't see the variable and can't warn it's undefined. The fix here is to also run RefineOperationVariables on the codegen context.

Reviewed By: jstejada, kassens

Differential Revision: D15756749

fbshipit-source-id: b0b2401662fd38c65f5ac7913e7165387f0fd62a