Fix for ConnectionHandler when field is unset
Summary: Fixes an edge-case where if a client connection *field* was unset, but the client connection record itself existed, pagination would fatal since it would try to create the client connection record (which already exists). I added a regression test to confirm the bug and fixed: in that case, we now update the parent record to point to the existing client record and merge new items into that record.

Reviewed By: jstejada

Differential Revision: D15765854

fbshipit-source-id: 9b0c7dde3eeda5327f075ba52360c79676914a5d