Fix MockResolvers for object in plural fiels
Summary:
Correctly handle MockResolvers for items in the plural fields by moving _getDefaultValuesForObject inside of the generateDataForField function.

So, what happened before.
- If we have a default value for an object inside a plural field - this value is created outside of the `generateDataForField`. And this value isn't an Array. So this line makes it so, we just ignored it here:

```
Array.isArray(defaults) ? defaults : Array(1).fill(),

```

So, `possibleDefaultValue` in this case is `undefined`.

If we move the generation of the possible default value for an object inside this function it fixes the problem. An object will be created for each item in the list.

Reviewed By: josephsavona

Differential Revision: D15859388

fbshipit-source-id: a06ac134c2bda8fa6f28231d7825044f0ca0f347