New Connection Model 3/n: generateAndCompile support for requiring modules
Summary:
NOTE: See D15891507 for more context on the new connection model

This lays more groundwork for the new connection model. The idea is that the user can provide a connection resolver that will be invoked during various lifecycle stages such as DataChecker, GC, normalization, etc. We already support generated nodes containing require() calls of other modules, but only for files that go through the actual compiler - notably, this doesn't work for `generateAndCompiler()` that we use in runtime unit tests. This diff changes `generateAndCompile()` to support passing a module map, rewriting nodes with CodeMarker strings to the corresponding module value.

This also adds a preliminary unit test that confirms a) the above works and b) that using `connection_resolver` does in  fact result in an "unsupported" invariant error.

Reviewed By: alunyov

Differential Revision: D15912740

fbshipit-source-id: 28aea71ba407ffc541aa099f4313a9c25f2897e8