remove dev only operation text for persisted operations
Summary:
When we persist an operation, we used to additionally write the full text as a dev only property for debugging purposes.

Since the query text can grow pretty large, it can be annoying when looking at diffs and probably has a not-insignificant impact when in development mode.

For devtools when it's interesting to view the expanded operation text, I think we should find a way to link from devtools to GraphiQL in a way that expands the text.

Reviewed By: alunyov

Differential Revision: D16159497

fbshipit-source-id: c3dc91fb5e24d7aa81d01f977e81527f959fd7b8